+++ tests/testthat/test-etl.R
@@ -9,7 +9,7 @@ test_that("sqlite works", {
   expect_true(file.exists(find_schema(cars_sqlite)))
   expect_message(find_schema(cars_sqlite, "my_crazy_schema", "etl"))
   expect_output(summary(cars_sqlite), "files")
-  expect_message(cars_sqlite %>% etl_create(), "success")
+  expect_message(cars_sqlite %>% etl_create(), "Uploading")
   expect_message(cars_sqlite %>% etl_init(), "Loading SQL script")
   expect_message(
     cars_sqlite %>% etl_cleanup(delete_raw = TRUE, delete_load = TRUE),
@@ -18,7 +18,7 @@ test_that("sqlite works", {
 
 test_that("dplyr works", {
   expect_message(cars <- etl("mtcars") %>%
-    etl_create(), regexp = "success")
+    etl_create(), regexp = "Uploading")
   expect_gt(length(src_tbls(cars)), 0)
   tbl_cars <- cars %>%
      tbl("mtcars")
@@ -30,7 +30,7 @@ test_that("dplyr works", {
   # double up the data
   expect_message(
     cars %>%
-      etl_update(), regexp = "success")
+      etl_update(), regexp = "Uploading")
   res2 <- tbl_cars %>%
     collect()
   expect_equal(nrow(res2), 2 * nrow(mtcars))